backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
cryptography==43.0.1
PyJWT==2.9.0
pytest==8.3.3
autogen-agentchat==0.2.40
pyautogen==0.2.34
